DDF-4463 Rename search-form-all-* to search-form-*
