+++ ui/packages/catalog-ui-search/src/main/webapp/component/dropdown/search-form-selector/dropdown.search-form-selector.view.js
@@ -20,7 +20,7 @@ const DropdownView = require('../dropdown.view')
 const template = require('./dropdown.search-form-selector.hbs')
 const SearchFormsList = require('../../search-form-list/search-form-list.view')
 const store = require('../../../js/store.js')
-const SearchFormCollection = require('../../search-form/search-form-all-collection-instance')
+const SearchFormCollection = require('../../search-form/search-form-collection-instance')
 const Backbone = require('backbone')
 
 module.exports = DropdownView.extend({
