+++ ui/packages/catalog-ui-search/src/main/webapp/component/search-form/search-form-tab-container.view.js
@@ -16,7 +16,7 @@
 const Marionette = require('marionette')
 const template = require('./search-form.collection.hbs')
 const SearchFormCollectionView = require('./search-form.collection.view')
-const SearchFormCollection = require('./search-form-all-collection-instance')
+const SearchFormCollection = require('./search-form-collection-instance')
 const CustomElements = require('../../js/CustomElements.js')
 const LoadingCompanionView = require('../loading-companion/loading-companion.view.js')
 const user = require('../singletons/user-instance')
