+++ ui/packages/catalog-ui-search/src/main/webapp/component/search-form/search-form.collection.js
